axfs_devfs

Struct DeviceFileSystem

pub struct DeviceFileSystem { /* private fields */ }

A device filesystem that manages device nodes.

This filesystem provides access to special device files similar to /dev directory in Unix-like systems. Devices must be registered at filesystem creation time.

Fields

parent - The parent filesystem mount point

root - The root directory containing device nodes

Implementations

impl DeviceFileSystem

pub fn new() -> Self

Creates a new device filesystem instance.

Returns

A new device filesystem with an empty root directory.

pub fn mkdir(&self, name: &'static str) -> Arc<DirNode>

Creates a subdirectory at the root directory.

This method creates a new directory node and adds it to the root.

Arguments

name - The name of the subdirectory to create

Returns

A reference to the created directory node.

pub fn add(&self, name: &'static str, node: VfsNodeRef)

Adds a device node to the root directory.

This method registers a device node in the filesystem. The node must implement the VFS node operations trait.

Arguments

name - The name of the device node

node - The device node reference to add